ANDROID: GKI: fix allmodconfig x86 build for vmwgfx drm driver
In commit 7db6c88bb52f ("drm/vmwgfx: Add seqno waiter for sync_files"),
a variable is defined in the middle of a {} scope, which is fine for
newer kernel versions, but for android13-5.10 we still warn about stuff
like this, which will break the build.  Fix this up by declaring the
variable in the proper place.

Note, this is only needed to fix the x86 allmodconfig build, no real
Android GKI system uses this driver in this branch.

Fixes: 7db6c88bb52f ("drm/vmwgfx: Add seqno waiter for sync_files")
Change-Id: Ie95f8fa4fd6b4ec3bc9ec04866598fd491af7f04
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>