mm: Prevent Kcompressd from breaking KMI [2/2]
The original commit added several extra members to struct pglist_data,
which KMI cannot accept.

Let's add an additional struct array kcompress_data containing various
fields of kcompressd. Its length is the same as node_data and it is
used in the same way as node_data (in fact, since GKI does not enable
NUMA, the length of this array is 1).

Change-Id: I5e784d754c7d523d4c7a0d36623bf884293b67f8
Signed-off-by: Pzqqt <821026875@qq.com>